+++ jeap-message-contract-messagetype-registry/src/main/java/ch/admin/bit/jeap/messagecontract/messagetype/repository/MessageTypeRepositoryConfiguration.java
@@ -1,11 +1,21 @@
 package ch.admin.bit.jeap.messagecontract.messagetype.repository;
 
+import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 
 @AutoConfiguration
 @ComponentScan
 @EnableConfigurationProperties(MessageTypeRepositoryProperties.class)
 public class MessageTypeRepositoryConfiguration {
+
+    @Bean
+    @ConditionalOnMissingBean(MeterRegistry.class)
+    MeterRegistry meterRegistry() {
+        return new SimpleMeterRegistry();
+    }
 }