+++ jeap-message-contract-messagetype-registry/src/main/java/ch/admin/bit/jeap/messagecontract/messagetype/repository/github/GitHubMessageTypeRepository.java
@@ -1,17 +1,19 @@
 package ch.admin.bit.jeap.messagecontract.messagetype.repository.github;
 
 import ch.admin.bit.jeap.messagecontract.messagetype.repository.MessageTypeRepository;
+import io.micrometer.core.instrument.MeterRegistry;
 
 import java.util.Map;
 
 public class GitHubMessageTypeRepository extends MessageTypeRepository {
 
-    public GitHubMessageTypeRepository(String gitUri, Map<String, String> parameters) {
+    public GitHubMessageTypeRepository(String gitUri, Map<String, String> parameters, MeterRegistry meterRegistry) {
         super(gitUri);
         assertRequiredParameters(parameters);
         super.setCredentialsProvider(new GitHubAppCredentialsProvider(
                 parameters.get("GITHUB_APP_ID"),
-                parameters.get("GITHUB_PRIVATE_KEY_PEM")
+                parameters.get("GITHUB_PRIVATE_KEY_PEM"),
+                meterRegistry
         ));
     }